Correctly parse numthreads as a long.

+++ java_pi_dir/pi.java
@@ -6,7 +6,7 @@ public class pi {
         numsteps = 500000000;
 
         if (argc.length > 0) {            
-            numsteps = Integer.parseInt(argc[0]);
+            numsteps = Long.parseLong(argc[0]);
         }
         
         System.out.println("Calculating PI using:");
